add protoencoder for metric
